refactor: change package name
package name changed `it.dohyun.nms.api.type.*` to `it.dohyun.nms.type.*`
