+++ v1_16_R3/src/main/java/it/dohyun/nms/nms/v1_16_R3/NMS.kt
@@ -1,7 +1,7 @@
 package it.dohyun.nms.nms.v1_16_R3
 
 import it.dohyun.nms.api.API
-import it.dohyun.nms.api.type.nms.NMS
+import it.dohyun.nms.type.nms.NMS
 import net.minecraft.server.v1_16_R3.IChatBaseComponent
 import net.minecraft.server.v1_16_R3.Packet
 import net.minecraft.server.v1_16_R3.PacketPlayOutPlayerListHeaderFooter
